Normalized Tsec with period of inner most planet
This must be done in order to allow time scaling consistency for systems that operate on different time scales. This does not affect any data, training, or analysis done to this point because all training/testing systems are set up so that the period of the inner most planet is one time step. By normalizing the Tsec we make it so that other systems will be consistent with our training/testing set. This also ensures the right integration time.